# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/file/download": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "file-api"
                ],
                "summary": "下载",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt凭证",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文件Id",
                        "name": "fileId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/file/initialUpload": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file-api"
                ],
                "summary": "初始化分片上传",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt凭证",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "reqBody",
                        "name": "reqBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.File_InitialUploadReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.File_InitialUploadRsp"
                        }
                    }
                }
            }
        },
        "/api/file/mergePart": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "file-api"
                ],
                "summary": "合并分片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt凭证",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "文件Id",
                        "name": "fileId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/file/uploadPart": {
            "patch": {
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "file-api"
                ],
                "summary": "上传分片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt凭证",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分片序号",
                        "name": "fileId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "文件Id",
                        "name": "chunkNum",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "protocol.File_InitialUploadReq": {
            "type": "object",
            "properties": {
                "md5": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sha1": {
                    "type": "string"
                },
                "sha256": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "protocol.File_InitialUploadRsp": {
            "type": "object",
            "properties": {
                "exists": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                }
            }
        }
    }
}